@@ -4,7 +4,7 @@ class ForgotPasswordsController < ApplicationController
   def create
     user = User.find_by(email: params[:email])
     if user
-      AppMailer.send_forgot_password(user).deliver
+      AppMailer.send_forgot_password(user).deliver_later
       redirect_to forgot_password_confirmation_path
     else
       flash[:error] = params[:email].blank? ? "Email can't be blank" : 'There is no user with that email in the system'
