@@ -8,7 +8,7 @@ class InvitationsController < ApplicationController
   def create
     @invitation = Invitation.new(invitation_params.merge!(inviter_id: current_user.id))
     if @invitation.save
-      AppMailer.send_invitation_email(@invitation).deliver
+      AppMailer.send_invitation_email(@invitation).deliver_later
       flash[:success] = "You have successfully invited #{@invitation.recipient_name}"
       redirect_to new_invitation_path
     else
